+++ src/common/js/configure.js
@@ -24,10 +24,21 @@ AppPages._defaults = {
  */
 AppPages.configure = function( o ){
     if( o && _.isObject( o )){
-        _conf = _.merge( AppPages._defaults, _conf, o );
-        AppPages._conf.set( _conf );
-        // be verbose if asked for
-        _verbose( AppPages.C.Verbose.CONFIGURE, 'configure() with', o );
+        // check that keys exist
+        let built_conf = {};
+        Object.keys( o ).forEach(( it ) => {
+            if( Object.keys( AppPages._defaults ).includes( it )){
+                built_conf[it] = o[it];
+            } else {
+                console.warn( 'pwix:app-pages configure() ignore unmanaged key \''+it+'\'' );
+            }
+        });
+        if( Object.keys( built_conf ).length ){
+            _conf = _.merge( AppPages._defaults, _conf, built_conf );
+            AppPages._conf.set( _conf );
+            // be verbose if asked for
+            _verbose( AppPages.C.Verbose.CONFIGURE, 'configure() with', built_conf );
+        }
     }
     // also acts as a getter
     return AppPages._conf.get();