+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - No more provide a default class to the display units
     - Remove useless RunContext.dataContext() method
     - Define new allowedInMenu() function, thus bumping minor candidate version number + enumerate() becomes async
+    - configure() now warns for unmanaged keys
 
 ### 1.2.0